#95 refactor: util 폴더 수정
