--- src/page/archiving/index/util/formattingDate.ts
@@ -1,7 +1,0 @@
-import { format } from 'date-fns';
-
-export const formattingDate = (date: Date) => {
-  const formattedDate = format(date, 'yyyy.MM.dd');
-
-  return formattedDate;
-};
