--- src/shared/util/getRandomColor.ts
@@ -1,4 +1,0 @@
-export const getRandomColor = (colors: string[]): string => {
-  const randomIndex = Math.floor(Math.random() * colors.length);
-  return colors[randomIndex];
-};
